"ads1_m14_a1: 4/4 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ Assignments/M14/14.1 Dynamic Median Finding/Solution.java
@@ -9,12 +9,12 @@ public class Solution {
         Scanner scan = new Scanner(System.in);
         int n = scan.nextInt();
 
-        MinPQ<Integer> minpq = new MinPQ<Integer>(n);
-        MaxPQ<Integer> maxpq = new MaxPQ<Integer>(n);
+        MinPQ<Float> minpq = new MinPQ<Float>(n);
+        MaxPQ<Float> maxpq = new MaxPQ<Float>(n);
 
-        int median = 0;
+        Float median = 0.0f;
         for (int i = 0; i < n; i++) {
-            int val = scan.nextInt();
+            Float val = scan.nextFloat();
             if (val > median) {
                 minpq.insert(val);
             } else {
